requirements.txt
asgiref==3.7.2
asn1crypto==1.3.0
attrs==18.2.0
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2

cssselect2==0.2.1
defusedxml==0.6.0
Deprecated==1.2.13
Django==3.2.19
django-cors-headers==4.0.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
idna==2.8
jdcal==1.4.1
olefile==0.46
openpyxl==3.0.3
ordered-set==4.1.0
pdfkit==0.6.1
pdflatex==0.1.3
psycopg2-binary==2.9.6
pycosat==0.6.3
pycparser==2.19
Pygments==2.6.1
PyJWT==1.7.1
pynput==1.6.8
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
python-xlib==0.33
pytz==2023.3
PyYAML==5.3.1
requests==2.22.0
ruamel-yaml
secure-smtplib==0.1.1
six==1.14.0
sqlparse==0.4.4
svgwrite==1.4
tinycss2==1.0.2
tqdm==4.42.1
typing-extensions==4.6.3
urllib3==1.25.8
webencodings==0.5.1
wrapt==1.15.0
XlsxWriter==1.2.8
